socialSound_definitiva/requirements.txt
Django~=5.1.1
django-seed~=0.3.1
Pillow~=10.4.0
django-debug-toolbar~=4.4.6
django-bootstrap5~=24.3
django-bootstrap-icons~=0.8.6
mutagen>=1.47.0
django-environ~=0.11.2
python-dotenv~=1.0.1
djangorestframework~=3.14.0
